Fix some minor issues and prepare for deploying to glassfish

+++ trunk/lib/video_processing.rb
@@ -5,13 +5,14 @@ module VideoProcessing
     if(File.exist?(converted_file_path))
       File.delete(converted_file_path)
     end
-    system("ffmpeg -i #{file_path} -acodec mp3 -ar 22050 -ab 32 -b 400 -f flv -s #{FLASH_VIDEO_RESOLUTION} #{converted_file_path}")
+    raise "ffmpeg command is unavailable at #{FFMPEG_PATH}, please define it in FFMPEG_PATH in environment.rb." unless File.exist?(FFMPEG_PATH)
+    system("#{FFMPEG_PATH} -i #{file_path} -acodec #{MP3_CODEC} -ar 22050 -ab 32 -b 400 -f flv -s #{FLASH_VIDEO_RESOLUTION} #{converted_file_path}")
     video_preview_image = flash_video_preview_image(file_path)
     if(File.exist?(video_preview_image))
       File.delete(video_preview_image)
     end
     # create a preview image for this video
-    system("ffmpeg -i #{converted_file_path} -vframes 1 -ss 00:00:03 -an -f mjpeg #{video_preview_image}")
+    system("#{FFMPEG_PATH} -i #{converted_file_path} -vframes 1 -ss 00:00:03 -an -f mjpeg #{video_preview_image}")
   end
 
   def flash_video_url(object, method)
